Move raw_val module to val
